Create CI-CD.yml based upon the workflow
created an initial version of our workflow. The logic behind is that I want to deploy CI CD PIpeline using github actions in such a way that when i push a new code or run a new model instead of a logistic regression model it runs smoothly. If i run a bad model or push a bad code it doesn't allow me to commit.